requirements.txt
Django>=5.2,<6.0
mysqlclient>=2.2.0
xhtml2pdf>=0.2.10
reportlab>=4.0.0
Pillow>=10.0.0
django-crispy-forms>=2.0
django-bootstrap5>=23.3
six>=1.16.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0 